Created a new directory for chapter 1
This new directory called psitn-1 houses files created during the
first chapter of the Pluralsight Introduction to Node.js training
course. From here on I will try to remember to create directories
for each of the chapters of the different learning resources so I
can keep up with where all of this stuff comes from.